Put textForm into new UserInputContainer.dart and cleaned code
